+++ tests/CouchDB.Driver.UnitTests/SupportByCombination_Tests.cs
@@ -14,7 +14,7 @@ namespace CouchDB.Driver.UnitTests
         private readonly CouchDatabase<Rebel> _rebels;
         private readonly Rebel _mainRebel;
         private readonly List<Rebel> _rebelsList;
-        private object _response;
+        private readonly object _response;
 
         public SupportByCombination_Tests()
         {
@@ -38,7 +38,7 @@ namespace CouchDB.Driver.UnitTests
         }
 
         [Fact]
-        public async Task Max()
+        public void Max()
         {
             using (var httpTest = new HttpTest())
             {
@@ -49,7 +49,7 @@ namespace CouchDB.Driver.UnitTests
         }
 
         [Fact]
-        public async Task Min()
+        public void Min()
         {
             using (var httpTest = new HttpTest())
             {
@@ -60,7 +60,7 @@ namespace CouchDB.Driver.UnitTests
         }
 
         [Fact]
-        public async Task First()
+        public void First()
         {
             using (var httpTest = new HttpTest())
             {
@@ -71,7 +71,7 @@ namespace CouchDB.Driver.UnitTests
         }
 
         [Fact]
-        public async Task First_Expr()
+        public void First_Expr()
         {
             using (var httpTest = new HttpTest())
             {
@@ -82,7 +82,7 @@ namespace CouchDB.Driver.UnitTests
         }
 
         [Fact]
-        public async Task FirstOrDefault()
+        public void FirstOrDefault()
         {
             using (var httpTest = new HttpTest())
             {
@@ -93,12 +93,74 @@ namespace CouchDB.Driver.UnitTests
         }
 
         [Fact]
-        public async Task FirstOrDefault_Expr()
+        public void FirstOrDefault_Expr()
         {
             using (var httpTest = new HttpTest())
             {
                 httpTest.RespondWithJson(new { Docs = Array.Empty<Rebel>() });
                 var result = _rebels.AsQueryable().FirstOrDefault(r => r.Age == 20);
+                Assert.Null(result);
+            }
+        }
+
+        [Fact]
+        public void Single()
+        {
+            using (var httpTest = new HttpTest())
+            {
+                httpTest.RespondWithJson(_response);
+                var result = _rebels.AsQueryable().Single();
+                Assert.Equal(_mainRebel.Age, result.Age);
+            }
+        }
+
+        [Fact]
+        public void Single_Exception()
+        {
+            using (var httpTest = new HttpTest())
+            {
+                httpTest.RespondWithJson(_response);
+                var result = _rebels.AsQueryable().Single();
+                Assert.Equal(_mainRebel.Age, result.Age);
+            }
+        }
+
+        [Fact]
+        public void Single_Expr()
+        {
+            using (var httpTest = new HttpTest())
+            {
+                httpTest.RespondWithJson(new
+                {
+                    Docs = new List<Rebel>
+                    {
+                        new Rebel(),
+                        new Rebel()
+                    }
+                });
+                var ex = Assert.Throws<InvalidOperationException>(() => _rebels.AsQueryable().Single());
+                Assert.Equal("Sequence contains more than one element", ex.Message);
+            }
+        }
+
+        [Fact]
+        public void SingleOrDefault()
+        {
+            using (var httpTest = new HttpTest())
+            {
+                httpTest.RespondWithJson(new { Docs = Array.Empty<Rebel>() });
+                var result = _rebels.AsQueryable().SingleOrDefault();
+                Assert.Null(result);
+            }
+        }
+
+        [Fact]
+        public void SingleOrDefault_Expr()
+        {
+            using (var httpTest = new HttpTest())
+            {
+                httpTest.RespondWithJson(new { Docs = Array.Empty<Rebel>() });
+                var result = _rebels.AsQueryable().SingleOrDefault(r => r.Age == 20);
                 Assert.Null(result);
             }
         }