+++ tests/CouchDB.Driver.UnitTests/_Models/Battle.cs
@@ -6,7 +6,9 @@ namespace CouchDB.Driver.UnitTests.Models
 {
     public class Battle
     {
+        public bool DidWin { get; set; }
         public string Planet { get; set; }
         public DateTime Date { get; set; }
+        public List<Vehicle> Vehicles { get; set; } = new List<Vehicle>();
     }
 }