+++ tests/CouchDB.Driver.UnitTests/_Models/Vehicle.cs
@@ -1,0 +1,7 @@
+﻿namespace CouchDB.Driver.UnitTests.Models
+{
+    public class Vehicle
+    {
+        public bool CanFly { get; set; }
+    }
+}