+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-﻿# 1.1.0 (2019-05-05)
+﻿# 1.1.1 (2019-06-02)
+
+## Features
+* **Single/SingleOrDefault:** Methods implementated as composite supported methods (Where and Take(2)).
+
+## Bug Fixes
+* **Queries:** Implicit bools in nested methods. ([#PR41](https://github.com/matteobortolazzo/couchdb-net/pull/41))
+* **FxCopAnalyzers:** Removed from NuGet dependencies.
+
+# 1.1.0 (2019-05-05)
 
 ## Features
 * **_find:** IQueryable methods that are not supported by CouchDB are evaluated in-memory using the IEnumerable counterpart, if possible.